🎨Update: admin page proptypes && 🚚Move: Constants to src from components
