+++ src/utils/socket.js
@@ -1,5 +1,5 @@
 import io from "socket.io-client";
-import CLIENT_ENDPOINT from "../components/constants";
+import CLIENT_ENDPOINT from "../constants/index";
 
 export const socket = io.connect(CLIENT_ENDPOINT, { cors: { origin: "*" } });
